Rename varen files and move config config from thema to config file
