+++ src/client/pages/Varen/Varen-thema-config.ts
@@ -1,8 +1,9 @@
 import { generatePath } from 'react-router-dom';
 
+import { VarenFrontend } from '../../../server/services/varen/config-and-types';
 import { AppRoutes } from '../../../universal/config/routes';
 import { dateSort } from '../../../universal/helpers/date';
-import { VarenFrontend } from '../../../server/services/varen/config-and-types';
+import { LinkProps } from '../../../universal/types';
 
 export const displayPropsAanvragen = {
   detailLinkComponent: '',
@@ -39,3 +40,17 @@ export const tableConfig = {
     ...tableConfigBase,
   },
 };
+
+export const linkListItems: LinkProps[] = [
+  {
+    to: 'https://www.amsterdam.nl/verkeer-vervoer/varen-amsterdam/varen-beroepsvaart/#:~:text=De%20passagiersvaart%20in%20Amsterdam%20is,stad%20willen%20we%20graag%20behouden.',
+    title: 'Meer informatie over passagiers- en beroepsvaart',
+  },
+];
+
+export const buttonItems: LinkProps[] = [
+  {
+    to: 'https://formulieren.acc.amsterdam.nl/TriplEforms/DirectRegelen/formulier/nl-NL/evAmsterdam/VARExploitatievergunningAanvragen.aspx',
+    title: 'Exploitatievergunning aanvragen',
+  },
+];
